+++ common/src/main/java/net/sashiro/compressedblocks/util/StringUtils.java
@@ -1,5 +1,10 @@
 package net.sashiro.compressedblocks.util;
 
+import java.io.IOException;
+import java.nio.file.Files;
+import java.nio.file.Path;
+import java.nio.file.StandardOpenOption;
+
 public class StringUtils {
     
     /**
@@ -19,5 +24,17 @@ public class StringUtils {
             stringFormat.append(first.toUpperCase()).append(afterFirst).append(" ");
         }
         return stringFormat.toString().trim();
+    }
+    
+    /**
+     * Helper function to write a {@link String} to a {@link java.io.File}
+     *
+     * @param content  the {@link String} to be written
+     * @param filePath the {@link Path} the {@link String} should be written to
+     * @throws IOException if {@link Path} not found / does not exist
+     */
+    public static void writeStringToFile(String content, Path filePath) throws IOException {
+        byte[] bytes = content.getBytes();
+        Files.write(filePath, bytes, StandardOpenOption.CREATE, StandardOpenOption.WRITE);
     }
 }